+++ src/plugins/grid.js
@@ -1,12 +1,15 @@
 bb.grid = {  
     apply: function(elements) {
 		if (bb.device.isBB10) {
-			var res;
-			if (bb.device.isPlayBook) {
-				res = 'lowres';
-			} else {
-				res = 'hires';
-			}
+			var res = (bb.device.isPlayBook) ? 'lowres' : 'hires',
+				R,G,B,
+				solidHeader = false,
+				headerJustify;
+
+			// Get our highlight RGB colors
+			R = parseInt((bb.slider.cutHex(bb.options.bb10HighlightColor)).substring(0,2),16)
+			G = parseInt((bb.slider.cutHex(bb.options.bb10HighlightColor)).substring(2,4),16);
+			B = parseInt((bb.slider.cutHex(bb.options.bb10HighlightColor)).substring(4,6),16);
 			// Apply our transforms to all grids
 			for (var i = 0; i < elements.length; i++) {
 				var j,
@@ -14,9 +17,23 @@ bb.grid = {
 					type,
 					title,
 					innerChildNode,
+					contextMenu,
 					outerElement = elements[i];
 					
-				outerElement.setAttribute('class','bb-bb10-grid-'+res);			
+				outerElement.setAttribute('class','bb-bb10-grid-'+res);	
+				// See if it is square or landscape layout
+				outerElement.isSquare = (outerElement.hasAttribute('data-bb-style') && outerElement.getAttribute('data-bb-style').toLowerCase() == 'square');
+				
+				// Get our header style
+				solidHeader = outerElement.hasAttribute('data-bb-header-style') ? (outerElement.getAttribute('data-bb-header-style').toLowerCase() == 'solid') : false;
+				// Get our header justification
+				headerJustify = outerElement.hasAttribute('data-bb-header-justify') ? outerElement.getAttribute('data-bb-header-justify').toLowerCase() : 'center';
+				
+				// Assign our context menu if there is one
+				if (outerElement.hasAttribute('data-bb-context') && outerElement.getAttribute('data-bb-context').toLowerCase() == 'true') {
+					contextMenu = bb.screen.contextMenu;
+				}
+				
 				// Gather our inner items
 				items = outerElement.querySelectorAll('[data-bb-type=group], [data-bb-type=row]');
 				for (j = 0; j < items.length; j++) {
@@ -26,16 +43,30 @@ bb.grid = {
 						type = innerChildNode.getAttribute('data-bb-type').toLowerCase();
 						if (type == 'group' && innerChildNode.hasAttribute('data-bb-title')) {
 							title = document.createElement('div');
-							title.normal = 'bb-bb10-grid-header-'+res+' bb10Accent';
-							title.highlight = 'bb-bb10-grid-header-'+res+' bb10Highlight';
+							title.normal = 'bb-bb10-grid-header-'+res;
 							title.innerHTML = '<p>'+ innerChildNode.getAttribute('data-bb-title') +'</p>';
+							
+							// Style our header for appearance
+							if (solidHeader) {
+								title.normal = title.normal +' bb10Accent';
+								title.style.color = 'white';
+								title.style['border-bottom-color'] = 'transparent';
+							} else {
+								title.normal = title.normal + ' bb-bb10-grid-header-normal-'+bb.screen.listColor;
+								title.style['border-bottom-color'] = 'rgb('+ (R - 32) +', '+ (G - 32) +', '+ (B - 32) +')';
+							}
+							
+							// Style our header for text justification
+							if (headerJustify == 'left') {
+								title.normal = title.normal + ' bb-bb10-grid-header-left-'+res;
+							} else if (headerJustify == 'right') {
+								title.normal = title.normal + ' bb-bb10-grid-header-right-'+res;
+							} else {
+								title.normal = title.normal + ' bb-bb10-grid-header-center';
+							}
+							
 							title.setAttribute('class', title.normal);
-							title.ontouchstart = function() {
-													this.setAttribute('class',this.highlight);
-												};
-							title.ontouchend = function() {
-													this.setAttribute('class',this.normal);
-												};
+							
 							if (innerChildNode.firstChild) {
 								innerChildNode.insertBefore(title, innerChildNode.firstChild);
 							} else {
@@ -53,6 +84,7 @@ bb.grid = {
 								subtitle,
 								height,
 								width,
+								hasOverlay,
 								rowItems = innerChildNode.querySelectorAll('[data-bb-type=item]');
 							
 							innerChildNode.setAttribute('class', 'bb-bb10-grid-row-'+res);
@@ -64,13 +96,19 @@ bb.grid = {
 							for (k = 0; k < numItems; k++) {
 								itemNode = rowItems[k];
 								subtitle = itemNode.innerHTML;
+								title = itemNode.getAttribute('data-bb-title');
+								hasOverlay = (subtitle || title);
 								itemNode.innerHTML = '';
 								if (bb.device.isPlayBook) {
 									width = ((window.innerWidth/numItems) - 5);
 								} else {
 									width = ((window.innerWidth/numItems) - 8);
 								}
-								height = Math.ceil(width*0.5625);
+								if (outerElement.isSquare) {
+									height = width;
+								} else {
+									height = Math.ceil(width*0.5625);
+								}
 								itemNode.setAttribute('class', 'bb-bb10-grid-item ' + columnClass);
 								itemNode.style.width = width + 'px';
 								itemNode.style.height = height + 'px';
@@ -81,20 +119,54 @@ bb.grid = {
 								image.setAttribute('style','height:100%;width:100%;');
 								itemNode.appendChild(image);
 								// Create our translucent overlay
-								overlay = document.createElement('div');
-								overlay.setAttribute('class','bb-bb10-grid-item-overlay-'+res);
-								overlay.innerHTML = '<div><p class="title">' + itemNode.getAttribute('data-bb-title') + '<br/>' + subtitle +'</p></div>';								
-								itemNode.appendChild(overlay);
-								// Add the overlay to the itemNode as a pointer for convenience when highlighting
-								itemNode.overlay = overlay;
-								itemNode.ontouchstart = function() {
-															this.overlay.setAttribute('style','opacity:1.0;background-color:' + bb.options.bb10HighlightColor +';');
-														};
-								itemNode.ontouchend = function() {
-															this.overlay.setAttribute('style','');
-														};
+								if (hasOverlay) {
+									overlay = document.createElement('div');
+									overlay.setAttribute('class','bb-bb10-grid-item-overlay-'+res);
+									overlay.innerHTML = '<div><p class="title">' + title + '<br/>' + subtitle +'</p></div>';								
+									itemNode.appendChild(overlay);
+									
+								} else {
+									overlay = null;
+								}
+								
 								itemNode.removeAttribute('data-bb-img');
 								itemNode.removeAttribute('data-bb-title');
+								
+								// Setup our variables
+								itemNode.overlay = overlay;
+								itemNode.title = title;
+								itemNode.description = subtitle;
+								itemNode.fingerDown = false;
+								itemNode.contextShown = false;
+								itemNode.contextMenu = contextMenu;
+								itemNode.ontouchstart = function() {
+															if (this.overlay) {
+																this.overlay.setAttribute('style','opacity:1.0;background-color:' + bb.options.bb10HighlightColor +';');
+															}
+															itemNode.fingerDown = true;
+															itemNode.contextShown = false;
+															if (itemNode.contextMenu) {
+																window.setTimeout(this.touchTimer, 667);
+															}
+														};
+								itemNode.ontouchend = function() {
+															if (this.overlay) {
+																this.overlay.setAttribute('style','');
+															}
+															itemNode.fingerDown = false;
+															if (itemNode.contextShown) {
+																event.preventDefault();
+																event.stopPropagation();
+															}
+														};
+								itemNode.touchTimer = function() {
+																if (itemNode.fingerDown) {
+																	itemNode.contextShown = true;
+																	itemNode.contextMenu.peek({title:this.title,description:this.description, selected: this});
+																}
+															};
+								itemNode.touchTimer = itemNode.touchTimer.bind(itemNode);
+								
 							}						
 							
 						}
@@ -126,7 +198,6 @@ bb.grid = {
 												innerWidth = 1280;
 											}
 										}
-										
 					
 										for (i = 0; i < items.length; i++) {
 											rowItems = items[i].querySelectorAll('[data-bb-type=item]');
@@ -138,7 +209,11 @@ bb.grid = {
 												} else {
 													width = ((innerWidth/numItems) - 8);
 												}
-												height = Math.ceil(width*0.5625);
+												if (outerElement.isSquare) {
+													height = width;
+												} else {
+													height = Math.ceil(width*0.5625);
+												}
 												itemNode.style.width = width+'px';
 												itemNode.style.height = height+'px';
 											}