+++ src/plugins/imageList.js
@@ -1,33 +1,55 @@
 bb.imageList = {  
     apply: function(elements) {
 		if (bb.device.isBB10) {
-			var res;
-			if (bb.device.isPlayBook) {
-				res = 'lowres';
-			} else {
-				res = 'hires';
-			}
+			var res = (bb.device.isPlayBook) ? 'lowres' : 'hires',
+				i,j,
+				outerElement,
+				innerChildNode,
+				normal,
+				highlight,
+				R,G,B,
+				contextMenu,
+				items,
+				hideImages,
+				imageEffect,
+				imagePlaceholder,
+				solidHeader = false,
+				headerJustify;
+				
+			// Get our highlight RGB colors
+			R = parseInt((bb.slider.cutHex(bb.options.bb10HighlightColor)).substring(0,2),16)
+			G = parseInt((bb.slider.cutHex(bb.options.bb10HighlightColor)).substring(2,4),16);
+			B = parseInt((bb.slider.cutHex(bb.options.bb10HighlightColor)).substring(4,6),16);
 		
 			// Apply our transforms to all Image Lists
-			for (var i = 0; i < elements.length; i++) {
-				var outerElement = elements[i],
-					normal,
-					highlight,
-					contextMenu;
+			for (i = 0; i < elements.length; i++) {
+				outerElement = elements[i];
 				outerElement.setAttribute('class','bb-bb10-image-list');
+				hideImages = outerElement.hasAttribute('data-bb-images') ? (outerElement.getAttribute('data-bb-images').toLowerCase() == 'none') : false;
+				if (!hideImages) {
+					imageEffect = outerElement.hasAttribute('data-bb-image-effect') ? outerElement.getAttribute('data-bb-image-effect').toLowerCase() : undefined;
+					imagePlaceholder = outerElement.hasAttribute('data-bb-image-placeholder') ? outerElement.getAttribute('data-bb-image-placeholder') : undefined;
+				}
+				
+				// Get our header style
+				solidHeader = outerElement.hasAttribute('data-bb-header-style') ? (outerElement.getAttribute('data-bb-header-style').toLowerCase() == 'solid') : false;
+				// Get our header justification
+				headerJustify = outerElement.hasAttribute('data-bb-header-justify') ? outerElement.getAttribute('data-bb-header-justify').toLowerCase() : 'center';
+				
 				// Assign our context menu if there is one
 				if (outerElement.hasAttribute('data-bb-context') && outerElement.getAttribute('data-bb-context').toLowerCase() == 'true') {
 					contextMenu = bb.screen.contextMenu;
 				}
 				// Gather our inner items
-				var items = outerElement.querySelectorAll('[data-bb-type=item], [data-bb-type=header]');
-				for (var j = 0; j < items.length; j++) {
-					var innerChildNode = items[j];
+				items = outerElement.querySelectorAll('[data-bb-type=item], [data-bb-type=header]');
+				for (j = 0; j < items.length; j++) {
+					innerChildNode = items[j];
 					if (innerChildNode.hasAttribute('data-bb-type')) {
 						// Figure out the type of item
 						var type = innerChildNode.getAttribute('data-bb-type').toLowerCase(),
 							description = innerChildNode.innerHTML,
 							title,
+							overlay,
 							accentText,
 							img,
 							details,
@@ -35,36 +57,29 @@ bb.imageList = {
 						
 						if (type == 'header') {
 							// Set our normal and highlight styling
-							normal = 'bb-bb10-image-list-header bb10Accent bb-bb10-image-list-header-'+res;
-							highlight = 'bb-bb10-image-list-header bb10Highlight bb-bb10-image-list-header-'+res;
+							normal = 'bb-bb10-image-list-header bb-bb10-image-list-header-'+res;
+							if (solidHeader) {
+								normal = normal +' bb10Accent';
+								innerChildNode.style.color = 'white';
+								title.style['border-bottom-color'] = 'transparent';
+							} else {
+								normal = normal + ' bb-bb10-image-list-header-normal-'+bb.screen.listColor;
+								innerChildNode.style['border-bottom-color'] = 'rgb('+ (R - 32) +', '+ (G - 32) +', '+ (B - 32) +')';
+							}
+							
 							// Check for alignment
-							if (innerChildNode.hasAttribute('data-bb-justify')) {
-								if (innerChildNode.getAttribute('data-bb-justify').toLowerCase() == 'left') {
-									normal = normal + ' bb-bb10-image-list-header-left-'+res;
-									highlight = highlight + ' bb-bb10-image-list-header-left-'+res;
-								} else if (innerChildNode.getAttribute('data-bb-justify').toLowerCase() == 'right') {
-									normal = normal + ' bb-bb10-image-list-header-right-'+res;
-									highlight = highlight + ' bb-bb10-image-list-header-right-'+res;
-								} else {
-									normal = normal + ' bb-bb10-image-list-header-center';
-									highlight = highlight + ' bb-bb10-image-list-header-center';
-								}
+							if (headerJustify == 'left') {
+								normal = normal + ' bb-bb10-image-list-header-left-'+res;
+							} else if (headerJustify == 'right') {
+								normal = normal + ' bb-bb10-image-list-header-right-'+res;
 							} else {
 								normal = normal + ' bb-bb10-image-list-header-center';
-								highlight = highlight + ' bb-bb10-image-list-header-center';
 							}
 							
 							// Set our styling
 							innerChildNode.normal = normal;
-							innerChildNode.highlight = highlight;
 							innerChildNode.innerHTML = '<p>'+ description +'</p>';
 							innerChildNode.setAttribute('class', normal);
-							innerChildNode.ontouchstart = function () {
-															this.setAttribute('class', this.highlight);
-														}
-							innerChildNode.ontouchend = function () {
-															this.setAttribute('class',this.normal);
-														}
 						}
 						else if (type == 'item') {
 							normal = 'bb-bb10-image-list-item bb-bb10-image-list-item-' + bb.screen.listColor + ' bb-bb10-image-list-item-' + res;
@@ -74,15 +89,56 @@ bb.imageList = {
 							innerChildNode.setAttribute('class', normal);
 							innerChildNode.innerHTML = '';
 							// Create our image
-							img = document.createElement('img');
-							img.setAttribute('src',innerChildNode.getAttribute('data-bb-img'));
-							innerChildNode.appendChild(img);
+							if (!hideImages) {
+								img = document.createElement('img');
+								if (imagePlaceholder) {
+									img.placeholder = imagePlaceholder;
+									img.src = innerChildNode.hasAttribute('data-bb-img') ? innerChildNode.getAttribute('data-bb-img') : imagePlaceholder;
+								} else {
+									img.setAttribute('src',innerChildNode.getAttribute('data-bb-img'));
+								}
+								innerChildNode.appendChild(img);
+								
+								if (imageEffect) {
+									img.style.opacity = '0.0';
+									img.even = (j%2 == 0)
+									img.onload = function() {
+													this.show();
+												};
+									img.show = function() {
+													this.style.opacity = '1.0';
+													if (this.even) { // Change timing based on even and odd numbers for some randomness
+														this.style['-webkit-transition'] = 'opacity 0.5s linear';
+													} else {
+														this.style['-webkit-transition'] = 'opacity 1.0s linear';
+													}
+													this.style['-webkit-backface-visibility'] = 'hidden';
+													this.style['-webkit-perspective'] = 1000;
+													this.style['-webkit-transform'] = 'translate3d(0,0,0)';
+												};
+									img.show = img.show.bind(img);
+								}
+								// Handle the error scenario
+								if (imagePlaceholder) {
+									img.onerror = function() {
+													if (this.src == this.placeholder) return;
+													this.src = this.placeholder;
+													if (imageEffect) {
+														this.show();
+													}
+												};
+								}
+							}
 							// Create the details container
 							details = document.createElement('div');
-							details.setAttribute('class','details');
+							if (hideImages) {
+								details.setAttribute('class','bb-bb10-image-list-item-details-'+res+' bb-bb10-image-list-item-noimage-'+res);
+							} else {
+								details.setAttribute('class','bb-bb10-image-list-item-details-'+res);
+							}
 							innerChildNode.appendChild(details);
 							// Create our title
-							title = document.createElement('span');
+							title = document.createElement('div');
 							title.setAttribute('class','title');
 							title.innerHTML = innerChildNode.getAttribute('data-bb-title');
 							details.appendChild(title);
@@ -101,14 +157,23 @@ bb.imageList = {
 							// Clean-up
 							innerChildNode.removeAttribute('data-bb-img');
 							innerChildNode.removeAttribute('data-bb-title');
+							// Add our highlight overlay
+							overlay = document.createElement('div');
+							overlay.setAttribute('class','bb-bb10-image-list-item-overlay-'+res);
+							innerChildNode.appendChild(overlay);
+								
 							// Set up our variables
 							innerChildNode.fingerDown = false;
 							innerChildNode.contextShown = false;
+							innerChildNode.overlay = overlay;
 							innerChildNode.contextMenu = contextMenu;
 							innerChildNode.description = description;
 							innerChildNode.title = title.innerHTML;
+							
+							
 							innerChildNode.ontouchstart = function () {
-															this.setAttribute('class',this.highlight);
+															//this.setAttribute('class',this.highlight);
+															this.overlay.style['border-color'] =  'rgb('+ (R - 32) +', '+ (G - 32) +', '+ (B - 32) +')';
 															innerChildNode.fingerDown = true;
 															innerChildNode.contextShown = false;
 															if (innerChildNode.contextMenu) {
@@ -116,7 +181,8 @@ bb.imageList = {
 															}
 														};
 							innerChildNode.ontouchend = function (event) {
-															this.setAttribute('class',this.normal);
+															//this.setAttribute('class',this.normal);
+															this.overlay.style['border-color'] = 'transparent';
 															innerChildNode.fingerDown = false;
 															if (innerChildNode.contextShown) {
 																event.preventDefault();
@@ -130,6 +196,23 @@ bb.imageList = {
 															}
 														};
 							innerChildNode.touchTimer = innerChildNode.touchTimer.bind(innerChildNode);
+							
+							// Add our subscription for click events to change highlighting on click
+							innerChildNode.trappedClick = innerChildNode.onclick;
+							innerChildNode.onclick = undefined;
+							innerChildNode.addEventListener('click',function (e) {
+									this.setAttribute('class',this.highlight);
+									if (this.trappedClick) {
+										setTimeout(this.trappedClick, 0);
+									}
+									setTimeout(this.finishHighlight, 250);
+								},false);
+								
+							// Finish the highlight on a delay
+							innerChildNode.finishHighlight = function() {
+														this.setAttribute('class',this.normal);
+													};
+							innerChildNode.finishHighlight = innerChildNode.finishHighlight.bind(innerChildNode);		
 						}
 					}
 				}
@@ -140,7 +223,14 @@ bb.imageList = {
 			for (var i = 0; i < elements.length; i++) {
 				var inEvent, 
 					outEvent, 
-					outerElement = elements[i];
+					outerElement = elements[i],
+					imagePlaceholder,
+					headerJustify,
+					hideImages = outerElement.hasAttribute('data-bb-images') ? (outerElement.getAttribute('data-bb-images').toLowerCase() == 'none') : false;
+					
+				if (!hideImages) {
+					imagePlaceholder = outerElement.hasAttribute('data-bb-image-placeholder') ? outerElement.getAttribute('data-bb-image-placeholder') : undefined;
+				}
 				// Set our highlight events
 				if (bb.device.isPlayBook) {
 					inEvent = 'ontouchstart';
@@ -154,6 +244,10 @@ bb.imageList = {
 				} else {
 					outerElement.setAttribute('class','bb-lowres-image-list');
 				}
+				
+				// Get our header justification
+				headerJustify = outerElement.hasAttribute('data-bb-header-justify') ? outerElement.getAttribute('data-bb-header-justify').toLowerCase() : 'center';
+				
 				// Gather our inner items
 				var items = outerElement.querySelectorAll('[data-bb-type=item], [data-bb-type=header]'),
 					innerChildNode,
@@ -163,13 +257,12 @@ bb.imageList = {
 					accentText,
 					normal,
 					highlight,
-					res;
-					
-				if (bb.device.isHiRes) {
-					res = 'hires';
-				} else {
-					res = 'lowres';
-				}
+					details,
+					titleDiv,
+					descriptionDiv,
+					accentDiv,
+					img,
+					res = (bb.device.isHiRes) ? 'hires' : 'lowres';
 					
 				for (j = 0; j < items.length; j++) {
 					innerChildNode = items[j];
@@ -187,17 +280,12 @@ bb.imageList = {
 							normal = 'bb-'+res+'-image-list-header';
 							highlight = 'bb-'+res+'-image-list-header-hover';
 							// Check for alignment
-							if (innerChildNode.hasAttribute('data-bb-justify')) {
-								if (innerChildNode.getAttribute('data-bb-justify').toLowerCase() == 'left') {
-									normal = normal + ' bb-'+res+'-image-list-header-left';
-									highlight = highlight + ' bb-'+res+'-image-list-header-left';
-								} else if (innerChildNode.getAttribute('data-bb-justify').toLowerCase() == 'right') {
-									normal = normal + ' bb-'+ res+'-image-list-header-right';
-									highlight = highlight + ' bb-'+res+'-image-list-header-right';
-								} else {
-									normal = normal + ' bb-'+res+'-image-list-header-center';
-									highlight = highlight + ' bb-'+res+'-image-list-header-center';
-								}
+							if (headerJustify == 'left') {
+								normal = normal + ' bb-'+res+'-image-list-header-left';
+								highlight = highlight + ' bb-'+res+'-image-list-header-left';
+							} else if (headerJustify == 'right') {
+								normal = normal + ' bb-'+ res+'-image-list-header-right';
+								highlight = highlight + ' bb-'+res+'-image-list-header-right';
 							} else {
 								normal = normal + ' bb-'+res+'-image-list-header-center';
 								highlight = highlight + ' bb-'+res+'-image-list-header-center';
@@ -212,16 +300,48 @@ bb.imageList = {
 							innerChildNode.setAttribute(outEvent, "this.setAttribute('class',this.normal)");
 						} 
 						else if (type == 'item') {
+							innerChildNode.innerHTML = '';
 							innerChildNode.setAttribute('class', 'bb-'+res+'-image-list-item');
 							innerChildNode.setAttribute(inEvent, "this.setAttribute('class','bb-"+res+"-image-list-item-hover')");
 							innerChildNode.setAttribute(outEvent, "this.setAttribute('class','bb-"+res+"-image-list-item')");
 							innerChildNode.setAttribute('x-blackberry-focusable','true');
-							innerChildNode.innerHTML = '<img src="'+ innerChildNode.getAttribute('data-bb-img') +'" />\n'+
-											'<div class="details">\n'+
-											'   <span class="title">' + innerChildNode.getAttribute('data-bb-title') + '</span>\n'+
-											'   <span class="accent-text">' + accentText + '</span>\n'+
-											'   <div class="description">' + description + '</div>\n'+
-											'</div>\n';
+							
+							if (!hideImages) {
+								img = document.createElement('img');
+								if (imagePlaceholder) {
+									img.placeholder = imagePlaceholder;
+									img.src = innerChildNode.hasAttribute('data-bb-img') ? innerChildNode.getAttribute('data-bb-img') : imagePlaceholder;
+									img.onerror = function() {
+													if (this.src == this.placeholder) return;
+													this.src = this.placeholder;
+												};
+								} else {
+									img.setAttribute('src',innerChildNode.getAttribute('data-bb-img') );
+								}
+								innerChildNode.appendChild(img);
+							}
+							
+							details = document.createElement('div');
+							innerChildNode.appendChild(details);
+							if (hideImages) {
+								details.setAttribute('class','bb-'+res+'-image-list-details bb-'+res+'-image-list-noimage');
+							} else {
+								details.setAttribute('class','bb-'+res+'-image-list-details');
+							}
+							
+							titleDiv = document.createElement('div');
+							titleDiv.innerHTML = innerChildNode.getAttribute('data-bb-title');
+							titleDiv.className = 'title';
+							details.appendChild(titleDiv);
+							accentDiv = document.createElement('div');
+							accentDiv.innerHTML = accentText;
+							accentDiv.className = 'accent-text';
+							details.appendChild(accentDiv);
+							descriptionDiv = document.createElement('div');
+							descriptionDiv.innerHTML = description;
+							descriptionDiv.className = 'description';
+							details.appendChild(descriptionDiv);
+							
 							innerChildNode.removeAttribute('data-bb-img');
 							innerChildNode.removeAttribute('data-bb-title');
 						}