+++ src/plugins/screen.js
@@ -42,24 +42,14 @@ bb.screen = {
 					scrollArea,
 					tempHolder = [],
 					childNode = null, 
-					j;
+					j,
+					height = (bb.device.isPlayBook) ? 73 : 140;
 				
 				// Figure out what to do with the title bar
                 if (titleBar.length > 0) {
 					titleBar = titleBar[0];
-					// See if they want a back button
-					if (titleBar.hasAttribute('data-bb-back-caption')) {
-						if (actionBar.length == 0) {
-							// Since there's no way to get back, we'll add an action bar
-							var newBackBar = document.createElement('div');
-							newBackBar.setAttribute('data-bb-type','action-bar');
-							newBackBar.setAttribute('data-bb-back-caption',titleBar.getAttribute('data-bb-back-caption'));
-							outerElement.appendChild(newBackBar);
-							actionBar = [newBackBar];
-						}
-					}
-					// TODO: Add title bar support
-					outerElement.removeChild(titleBar);
+				} else {
+					titleBar = null;
 				}
 				
 				// Assign our action bar
@@ -79,12 +69,12 @@ bb.screen = {
 				
 				// Inner Scroll Area
 				scrollArea = document.createElement('div');
-				outerScrollArea.appendChild(scrollArea); 
+				outerScrollArea.appendChild(scrollArea); 			
 				
 				// Copy all nodes in the screen that are not the action bar
 				for (j = 0; j < outerElement.childNodes.length - 1; j++) {
 					childNode = outerElement.childNodes[j];
-					if ((childNode != actionBar) && (childNode != menuBar)) {
+					if ((childNode != actionBar) && (childNode != menuBar) && (childNode != titleBar)) {
 						tempHolder.push(childNode);
 					}
 				}
@@ -93,17 +83,26 @@ bb.screen = {
 					scrollArea.appendChild(tempHolder[j]);
 				}
 				
-				if (actionBar) {
-					if (bb.device.isPlayBook) {
-						outerScrollArea.setAttribute('style','overflow:auto;position:absolute;bottom:73px;top:0px;left:0px;right:0px;');
-					} else {
-						outerScrollArea.setAttribute('style','overflow:auto;position:absolute;bottom:140px;top:0px;left:0px;right:0px;');
-					}
-					bb.actionBar.apply(actionBar,outerElement);
-                }
-				else {
+				// Set our outer scroll area dimensions
+				if (titleBar && actionBar) {
+					outerScrollArea.setAttribute('style','overflow:auto;position:absolute;bottom:'+height+'px;top:'+height+'px;left:0px;right:0px;');
+				} else if (titleBar) {
+					outerScrollArea.setAttribute('style','overflow:auto;bottom:0px;position:absolute;top:'+height+'px;left:0px;right:0px;');
+				} else if (actionBar) {
+					outerScrollArea.setAttribute('style','overflow:auto;position:absolute;bottom:'+height+'px;top:0px;left:0px;right:0px;');
+				} else {
 					outerScrollArea.setAttribute('style','overflow:auto;bottom:0px;position:absolute;top:0px;left:0px;right:0px;');
 				}
+				
+				// Apply any title bar styling
+				if (titleBar) {		
+					bb.titleBar.apply(titleBar);
+                }
+				
+				// Apply any action Bar styling
+				if (actionBar) {
+					bb.actionBar.apply(actionBar,outerElement);
+                }
 				
 				// Assign our context
 				if (context.length > 0) {
@@ -161,18 +160,7 @@ bb.screen = {
                    
 				if (titleBar) {
 					outerScrollArea.setAttribute('style','overflow:auto;bottom:0px;position:absolute;top:55px;left:0px;right:0px;');					
-                    titleBar.setAttribute('class', 'pb-title-bar');
-                    titleBar.innerHTML = titleBar.getAttribute('data-bb-caption');
-                    if (titleBar.hasAttribute('data-bb-back-caption')) {
-                        var button = document.createElement('div'), 
-                            buttonInner = document.createElement('div');
-                        button.setAttribute('class', 'pb-title-bar-back');
-                        button.onclick = bb.popScreen;
-                        buttonInner.setAttribute('class','pb-title-bar-back-inner');
-                        buttonInner.innerHTML = titleBar.getAttribute('data-bb-back-caption'); 
-                        button.appendChild(buttonInner);
-                        titleBar.appendChild(button);
-                    }
+                    bb.titleBar.apply(titleBar);
                 }
 				else {
 					outerScrollArea.setAttribute('style','overflow:auto;bottom:0px;position:absolute;top:0px;left:0px;right:0px;');
@@ -193,20 +181,9 @@ bb.screen = {
 					context[j].style.display = 'none';
 				}
 				
-				
                 if (titleBar.length > 0) {
                     titleBar = titleBar[0];
-                    if (titleBar.hasAttribute('data-bb-caption')) {
-                        var outerStyle = outerElement.getAttribute('style');
-                        if (bb.device.isHiRes) {
-                            titleBar.setAttribute('class', 'bb-hires-screen-title');
-                            outerElement.setAttribute('style', outerStyle + ';padding-top:33px');
-                        } else {
-                            titleBar.setAttribute('class', 'bb-lowres-screen-title');
-                            outerElement.setAttribute('style', outerStyle + ';padding-top:27px');
-                        }
-                        titleBar.innerHTML = titleBar.getAttribute('data-bb-caption');
-                    }
+                    bb.titleBar.apply(titleBar);
                 }
             }
         }